Refactored driver to active_driver, Added new files, Added a test case for simple login.
