+++ Data/output/output_data.py
@@ -1,0 +1,1 @@
+ERROR_MESSAGE = f"The password that you\'ve entered is incorrect.Forgotten password?"
