+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+
+markers =
+    Debug
+    Facebook
